Make all canrytokens.org links default to using https
Summary:
A user of `canarytokens.org mailed in with the following concern:

```
I was just testing out the DNS Canary Token. The canary triggered as expected but
the email alert I received didn't direct me to the secure version of canarytokens.org,
so my request (and therefore canary ID) was made unencrypted.
```

This is a very reasonable request, and we have thus now ensured that all links pointing
to canarytokens.org, including the token management and token history links in notifications
as well as a couple of informational/branding links on the site are using `https` instead of
just `http`. This included making a a similar fix to one or 2 `canary.tools` links as well.

Test Plan:
Ran the full docker dev setup locally on my machine, created and triggered both
http and dns tokens, and manually verified all links in the notification emails that were
sent as well as links on the site itself.

(See attached clip for more detail)

Reviewers: #devs

Differential Revision: http://phabricator.thinkst.com:9000/D2849